refactor: extract checkout + webhook routes out of the app factory (#107)
The revenue-critical routes move into backend/routes/ as injectable
router factories, each the Stripe adapter for its direction (ADR-0001).
The webhook router mounts before express.json() so signature
verification keeps the raw body; the checkout limiter moves with its
route. Behavior pinned by the integration tests in the previous commit.
